[refactor] Move SyncExchange changes to separate package under package server.
On second thoughts, we want a pluggable interface, such that we can fall
back to the HTTP implementation, just in case. Any ZMQ related changes
are better off in a separate package. This will help in make the
integration pluggable, although I don't quite know how at this point.
Lets figure it out.

- Related: #6155